fix search/save
