+++ 07-google-book-search/client/src/pages/Search.js
@@ -1,4 +1,4 @@
-import React, {useState} from "react";
+import React, { useState, useEffect } from "react";
 import Navbar from "../components/Navbar"
 import SearchBar from "../components/SearchBar"
 import Results from "../components/Results"
@@ -8,14 +8,25 @@ import "./style.css";
 
 
 function Search() {
-    
-    const [searchState, setSearchState]= useState("");
+
+    const [searchState, setSearchState] = useState("");
     const [books, setBooks] = useState([]);
     const [loading, setLoading] = useState(true);
+    const [modalClass, setModalClass] = useState("modal hideModal");
+    const [text, setText] = useState("Saved!")
 
-        
+    useEffect(() => {
+    }, [modalClass]);
 
-    
+    function modalClose() {
+        setModalClass("modal hideModal");
+    }
+
+    function saveModal(book) {
+        setModalClass("modal showModal");
+        setText(book.volumeInfo.title + " was saved!");
+    }
+
     const handleSearchChange = (e) => {
         const { value } = e.target
         setSearchState(value)
@@ -24,55 +35,52 @@ function Search() {
 
 
     const searchBooks = async () => {
-        
-       let newBooks = await APIbooks.getBooks(searchState)
-        .then((res) => {
-            setLoading(false);
-            return res.data.items;
-            console.log(books)
-        })
+
+        let newBooks = await APIbooks.getBooks(searchState)
+            .then((res) => {
+                setLoading(false);
+                return res.data.items;
+                console.log(books)
+            })
         console.log("newBooks: ", newBooks);
         setBooks(newBooks);
         setLoading(false);
     }
 
     const saveBook = (book) => {
-        var image;
-        if(book.volumeInfo.imageLinks===undefined){
-            image = "./googlebookslogo.png"
-        } else {
-            image = book.volumeInfo.imageLinks.thumbnail
+        const data = {
+            title: book.volumeInfo.title,
+            author: book.volumeInfo.authors,
+            description: book.volumeInfo.description,
+            image: book.volumeInfo.imageLinks.thumbnail,
+            link: book.volumeInfo.infoLink
         }
 
-        const data = {
-            title:book.volumeInfo.title,
-            author:book.volumeInfo.authors,
-            description:book.volumeInfo.description,
-            image: image,
-            link:book.volumeInfo.infoLink
-        }
-       
-        APIbooks.addBook(data).then(res=>{
-            console.log("saved",res)
-            
+        APIbooks.addBook(data).then(res => {
+            console.log("saved", res)
+
         });
     }
-    
- 
 
-    return(
-        <div className="mb-5">
-        <Navbar/>
-        <Jumbotron />
-        
-        <SearchBar 
-        handleSearchChange={handleSearchChange}
-        searchBooks={searchBooks}/>    
-        <Results
-        data={books}
-        saveBook={saveBook}
-        />
-          
+
+
+    return (
+        <div>
+            <Navbar />
+            <Jumbotron />
+
+            <SearchBar
+                handleSearchChange={handleSearchChange}
+                searchBooks={searchBooks} />
+            <Results
+                data={books}
+                saveBook={saveBook}
+                modalClose={modalClose}
+                saveModal={saveModal}
+                text={text}
+                modalClass={modalClass}
+            />
+
 
         </div>
     )
